move: product 컨트롤러 파일 위치 변경
